Integration tests refactoring
Split test utils: move tools and structured output examples to separate dirs, add separate APIKeys util
Improve test cases: add checks that tool was called, added some corner cases for anuOf and null args
