+++ integration-tests/src/jvmTest/kotlin/ai/koog/integration/tests/agent/AIAgentTestBase.kt
@@ -16,15 +16,15 @@ import ai.koog.agents.core.tools.ToolRegistry
 import ai.koog.agents.ext.agent.subgraphWithTask
 import ai.koog.agents.features.eventHandler.feature.EventHandler
 import ai.koog.agents.features.eventHandler.feature.EventHandlerConfig
+import ai.koog.integration.tests.utils.APIKeys.readTestAnthropicKeyFromEnv
+import ai.koog.integration.tests.utils.APIKeys.readTestOpenAIKeyFromEnv
 import ai.koog.integration.tests.utils.Models
-import ai.koog.integration.tests.utils.TestUtils.CreateFile
-import ai.koog.integration.tests.utils.TestUtils.DeleteFile
-import ai.koog.integration.tests.utils.TestUtils.ListFiles
-import ai.koog.integration.tests.utils.TestUtils.MockFileSystem
-import ai.koog.integration.tests.utils.TestUtils.ReadFile
-import ai.koog.integration.tests.utils.TestUtils.readTestAnthropicKeyFromEnv
-import ai.koog.integration.tests.utils.TestUtils.readTestOpenAIKeyFromEnv
 import ai.koog.integration.tests.utils.getLLMClientForProvider
+import ai.koog.integration.tests.utils.tools.files.CreateFile
+import ai.koog.integration.tests.utils.tools.files.DeleteFile
+import ai.koog.integration.tests.utils.tools.files.ListFiles
+import ai.koog.integration.tests.utils.tools.files.MockFileSystem
+import ai.koog.integration.tests.utils.tools.files.ReadFile
 import ai.koog.prompt.dsl.ModerationResult
 import ai.koog.prompt.dsl.Prompt
 import ai.koog.prompt.dsl.prompt
