+++ integration-tests/src/jvmTest/kotlin/ai/koog/integration/tests/agent/OllamaAgentIntegrationTest.kt
@@ -16,11 +16,11 @@ import ai.koog.agents.features.eventHandler.feature.EventHandler
 import ai.koog.integration.tests.InjectOllamaTestFixture
 import ai.koog.integration.tests.OllamaTestFixture
 import ai.koog.integration.tests.OllamaTestFixtureExtension
-import ai.koog.integration.tests.tools.AnswerVerificationTool
-import ai.koog.integration.tests.tools.GenericParameterTool
-import ai.koog.integration.tests.tools.GeographyQueryTool
 import ai.koog.integration.tests.utils.annotations.Retry
 import ai.koog.integration.tests.utils.annotations.RetryExtension
+import ai.koog.integration.tests.utils.tools.AnswerVerificationTool
+import ai.koog.integration.tests.utils.tools.GenericParameterTool
+import ai.koog.integration.tests.utils.tools.GeographyQueryTool
 import ai.koog.prompt.dsl.prompt
 import ai.koog.prompt.executor.model.PromptExecutor
 import ai.koog.prompt.llm.OllamaModels
