+++ integration-tests/src/jvmTest/kotlin/ai/koog/integration/tests/utils/APIKeys.kt
@@ -1,0 +1,45 @@
+package ai.koog.integration.tests.utils
+
+object APIKeys {
+    fun readTestAnthropicKeyFromEnv(): String {
+        return System.getenv("ANTHROPIC_API_TEST_KEY")
+            ?: error("ERROR: environment variable `ANTHROPIC_API_TEST_KEY` is not set")
+    }
+
+    fun readTestOpenAIKeyFromEnv(): String {
+        return System.getenv("OPEN_AI_API_TEST_KEY")
+            ?: error("ERROR: environment variable `OPEN_AI_API_TEST_KEY` is not set")
+    }
+
+    fun readTestGoogleAIKeyFromEnv(): String {
+        return System.getenv("GEMINI_API_TEST_KEY")
+            ?: error("ERROR: environment variable `GEMINI_API_TEST_KEY` is not set")
+    }
+
+    fun readTestOpenRouterKeyFromEnv(): String {
+        return System.getenv("OPEN_ROUTER_API_TEST_KEY")
+            ?: error("ERROR: environment variable `OPEN_ROUTER_API_TEST_KEY` is not set")
+    }
+
+    fun readTestMistralAiKeyFromEnv(): String {
+        return System.getenv("MISTRAL_AI_API_TEST_KEY")
+            ?: error("ERROR: environment variable `MISTRAL_AI_API_TEST_KEY` is not set")
+    }
+
+    fun readAwsAccessKeyIdFromEnv(): String {
+        return System.getenv("AWS_ACCESS_KEY_ID")
+            ?: error("ERROR: environment variable `AWS_ACCESS_KEY_ID` is not set")
+    }
+
+    fun readAwsSecretAccessKeyFromEnv(): String {
+        return System.getenv("AWS_SECRET_ACCESS_KEY")
+            ?: error("ERROR: environment variable `AWS_SECRET_ACCESS_KEY` is not set")
+    }
+
+    fun readAwsSessionTokenFromEnv(): String? {
+        return System.getenv("AWS_SESSION_TOKEN")
+            ?: null.also {
+                println("WARNING: environment variable `AWS_SESSION_TOKEN` is not set, using default session token")
+            }
+    }
+}
